test: Fold importVariables tests into the existing source-control suite
Move the new importVariables() test cases into the shared
SourceControlImportService integration suite instead of a standalone
file, matching how every other method on the service is already
tested there, and drop the Arrange/Act/Assert comments to match this
file's plain style.
